moved stuff in PersistStructs.hpp into EpochSys.hpp and Recoverable.hpp
